* Bump picomatch from 2.3.1 to 2.3.2 (#3111)

Bumps [picomatch](https://github.com/micromatch/picomatch) from 2.3.1 to 2.3.2.
- [Release notes](https://github.com/micromatch/picomatch/releases)
- [Changelog](https://github.com/micromatch/picomatch/blob/master/CHANGELOG.md)
- [Commits](https://github.com/micromatch/picomatch/compare/2.3.1...2.3.2)

---
updated-dependencies:
- dependency-name: picomatch
  dependency-version: 2.3.2
  dependency-type: indirect
...


* Bump yaml from 1.10.2 to 1.10.3 (#3110)

Bumps [yaml](https://github.com/eemeli/yaml) from 1.10.2 to 1.10.3.
- [Release notes](https://github.com/eemeli/yaml/releases)
- [Commits](https://github.com/eemeli/yaml/compare/v1.10.2...v1.10.3)

---
updated-dependencies:
- dependency-name: yaml
  dependency-version: 1.10.3
  dependency-type: indirect
...


* Additional checks on admin/instructor actions (#3113)

* Additional checks on admin actions.
* Adding additional checks for viewing specific data in admin tables.
* Adding current user can view for other admin tables.

* Adding attribution.

* Build for 9.2.3.

---------

+++ includes/controllers/class.llms.controller.quizzes.php
@@ -40,6 +40,7 @@ class LLMS_Controller_Quizzes {
 	 *
 	 * @since 3.37.8
 	 * @since 5.1.0 Use a deep orphan check to determine if the quiz can be deleted.
+	 * @since 9.2.3 Add capability check before deleting quiz.
 	 *
 	 * @return null|false|WP_Post `null` if the form wasn't submitted or the nonce couldn't be verified.
 	 *                            `false` if an error was encountered.
@@ -53,6 +54,9 @@ class LLMS_Controller_Quizzes {
 
 		$id = llms_filter_input( INPUT_POST, 'llms_del_quiz', FILTER_SANITIZE_NUMBER_INT );
 		if ( $id && 'llms_quiz' === get_post_type( $id ) ) {
+			if ( ! current_user_can( 'delete_post', $id ) ) {
+				return false;
+			}
 			$quiz = llms_get_post( $id );
 			if ( $quiz && ( $quiz->is_orphan( true ) || ! $quiz->get_course() ) ) {
 				return wp_delete_post( $id, true );
